requirements.txt
python-dotenv==1.0.0
google-cloud-pubsub==2.18.4
avro==1.11.3
protobuf==4.24.4
avro==1.11.3
